Added clamis to verify if is admin and added jwt_optional to /items

+++ resources/item.py
@@ -1,6 +1,7 @@
 # Python libraries
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims, \
+    jwt_optional, get_jwt_identity
 # Own libraries
 from models.item import ItemModel
 
@@ -49,12 +50,17 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims["is_admin"]:
+            return {"message": "Admin priviliges required"}
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
-
-        return {"message": "Item deleted"}
+            return {"message": "Item deleted"}
+        return {"message": "Item not found"}, 404
 
     def put(self, name):
         data = Item.parser.parse_args()
@@ -71,8 +77,16 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        # Lambda Function would be:
-        #   list(map(lambda item: item.json(), ItemModel.query.all()))
-        # List Comprehension is:
-        return {"items": [item.json() for item in ItemModel.query.all()]}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.query.all()]
+        if user_id:
+            # Lambda Function would be:
+            #   list(map(lambda item: item.json(), ItemModel.query.all()))
+            # List Comprehension is:
+            return {"items": items}, 200
+        return {
+            "items": [item["name"] for item in items],
+            "message": "More data available if you log in"
+        }, 200
